DIV-3300 - renamed RespondWithPin to CaptureCaseAndPin
