+++ config/default.yml
@@ -54,7 +54,7 @@ paths:
   health: /health
   authenticated: /authenticated
   index: /
-  respondWithPin: /respond-with-pin
+  captureCaseAndPin: /capture-case-and-pin
   respond: /respond
   reviewApplication: /review-application
   chooseAResponse: /choose-a-response
