+++ steps/capture-case-and-pin/CaptureCaseAndPin.content.json
