+++ steps/capture-case-and-pin/CaptureCaseAndPin.html
